Enhancement: Extract RateAction from TalksController
