+++ classes/Http/Controller/Admin/TalksController.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace OpenCFP\Http\Controller\Admin;
 
 use OpenCFP\Domain\Talk\TalkHandler;
-use OpenCFP\Domain\ValidationException;
 use OpenCFP\Http\Controller\BaseController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -34,23 +33,6 @@ class TalksController extends BaseController
         $this->talkHandler = $talkHandler;
 
         parent::__construct($twig, $urlGenerator);
-    }
-
-    public function rateAction(Request $request): Response
-    {
-        try {
-            $this->validate($request, [
-                'rating' => 'required|integer',
-            ]);
-
-            $content = (string) $this->talkHandler
-                ->grabTalk((int) $request->get('id'))
-                ->rate((int) $request->get('rating'));
-        } catch (ValidationException $e) {
-            $content = '';
-        }
-
-        return new Response($content);
     }
 
     /**
