+++ classes/Http/Controller/BaseController.php
@@ -13,12 +13,6 @@ declare(strict_types=1);
 
 namespace OpenCFP\Http\Controller;
 
-use Illuminate\Container\Container;
-use Illuminate\Filesystem\Filesystem;
-use Illuminate\Translation\FileLoader;
-use Illuminate\Translation\Translator;
-use Illuminate\Validation\Factory;
-use OpenCFP\Domain\ValidationException;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -108,23 +102,5 @@ abstract class BaseController
     protected function redirectBack(Request $request)
     {
         return new RedirectResponse($request->headers->get('referer'));
-    }
-
-    protected function validate(Request $request, $rules = [], $messages = [], $customAttributes = [])
-    {
-        $data = $request->query->all() + $request->request->all() + $request->files->all();
-
-        $validation = new Factory(
-            new Translator(new FileLoader(new Filesystem(), __DIR__ . '/../../../resources/lang'), 'en'),
-            new Container()
-        );
-
-        $validator = $validation->make($data, $rules, $messages, $customAttributes);
-
-        if ($validator->fails()) {
-            throw ValidationException::withErrors(array_flatten($validator->errors()->toArray()));
-        }
-
-        unset($validation, $validator);
     }
 }
