+++ tests/Integration/Http/Controller/Admin/TalksControllerTest.php
@@ -140,68 +140,4 @@ final class TalksControllerTest extends WebTestCase
         $this->assertResponseBodyNotContains('1', $response);
         ;
     }
-
-    /**
-     * @test
-     * @dataProvider providerValidRating
-     *
-     * @param mixed $rating
-     */
-    public function rateActionWorksCorrectly($rating)
-    {
-        $talk = self::$talks->first();
-
-        $response = $this
-            ->asAdmin()
-            ->post('/admin/talks/' . $talk->id . '/rate', [
-                'rating' => $rating,
-            ]);
-
-        $this->assertResponseIsSuccessful($response);
-        $this->assertResponseBodySame('1', $response);
-    }
-
-    public function providerValidRating(): array
-    {
-        return [
-            'int' => [
-                1,
-            ],
-            'integerish' => [
-                '0',
-            ],
-        ];
-    }
-
-    /**
-     * @test
-     * @dataProvider providerInvalidRating
-     *
-     * @param mixed $rating
-     */
-    public function rateActionReturnsFalseOnWrongRate($rating)
-    {
-        $talk = self::$talks->first();
-
-        $response = $this
-            ->asAdmin()
-            ->post('/admin/talks/' . $talk->id . '/rate', [
-                'rating' => $rating,
-            ]);
-
-        $this->assertResponseIsSuccessful($response);
-        $this->assertResponseBodyEmpty($response);
-    }
-
-    public function providerInvalidRating(): array
-    {
-        return [
-            'int-too-large' => [
-                12,
-            ],
-            'string' => [
-                'blabla',
-            ],
-        ];
-    }
 }
